use .antd-tools.config.js
