+++ tests/dekko/dist.test.js
@@ -1,12 +1,22 @@
 const $ = require('dekko');
 const chalk = require('chalk');
 
+$('lib/version')
+  .isDirectory()
+  .hasFile('index.js')
+  .hasFile('index.d.ts');
+
+$('lib/style')
+  .isDirectory()
+  .hasFile('components.less');
+
 $('dist')
   .isDirectory()
   .hasFile('antd.css')
   .hasFile('antd.min.css')
   .hasFile('antd.js')
-  .hasFile('antd.min.js');
+  .hasFile('antd.min.js')
+  .hasFile('antd.less');
 
 // eslint-disable-next-line
 console.log(chalk.green('✨ `dist` directory is valid.'));
